+++ prolif/interactions.py
@@ -43,6 +43,7 @@ from .utils import (
 import numpy as np
 from rdkit.Chem import MolFromSmarts
 from rdkit import Geometry
+from MDAnalysis.topology.tables import vdwradii
 
 _INTERACTIONS = {}
 
@@ -452,3 +453,42 @@ class MetalAcceptor(_BaseMetallic):
     def detect(self, ligand, residue):
         bit, ires, ilig = super().detect(residue, ligand)
         return bit, ilig, ires
+
+
+class VdWContact(Interaction):
+    """Interaction based on the van der Waals radii of interacting atoms.
+    
+    Parameters
+    ----------
+    tolerance : float
+        Tolerance added to the sum of vdW radii of atoms before comparing to
+        the interatomic distance. If ``distance <= sum_vdw + tolerance`` the
+        atoms are identified as a contact
+
+    Raises
+    ------
+    ValueError : ``tolerance`` parameter cannot be negative
+    """
+    def __init__(self, tolerance=.5):
+        if tolerance >= 0:
+            self.tolerance = tolerance
+        else:
+            raise ValueError("`tolerance` must be 0 or positive")
+        self._vdw_cache = {}
+
+    def detect(self, ligand, residue):
+        lxyz = ligand.GetConformer()
+        rxyz = residue.GetConformer()
+        for la, ra in product(ligand.GetAtoms(), residue.GetAtoms()):
+            lig = la.GetSymbol().upper()
+            res = ra.GetSymbol().upper()
+            try:
+                vdw = self._vdw_cache[(lig, res)]
+            except KeyError:
+                vdw = vdwradii[lig] + vdwradii[res] + self.tolerance
+                self._vdw_cache[(lig, res)] = vdw
+            dist = (lxyz.GetAtomPosition(la.GetIdx())
+                        .Distance(rxyz.GetAtomPosition(ra.GetIdx())))
+            if dist <= vdw:
+                return True, la.GetIdx(), ra.GetIdx()
+        return False, None, None